Move charts and collapsibles into separate dirs; fix css for download button
