+++ lib/Models/registerCustomComponentTypes.js
@@ -5,16 +5,16 @@
 const defined = require('terriajs-cesium/Source/Core/defined');
 const DeveloperError = require('terriajs-cesium/Source/Core/DeveloperError');
 
-const Chart = require('../ReactViews/Chart/Chart');
-const ChartExpandButton = require('../ReactViews/Chart/ChartExpandButton');
-const Collapsible = require('../ReactViews/Chart/Collapsible');
+const Chart = require('../ReactViews/Custom/Chart/Chart');
+const ChartExpandButton = require('../ReactViews/Custom/Chart/ChartExpandButton');
+const Collapsible = require('../ReactViews/Custom/Collapsible/Collapsible');
 const CustomComponents = require('./CustomComponents');
 const CustomComponentType = require('./CustomComponentType');
 const TableStructure = require('../Map/TableStructure');
 
 const React = require('react');
 
-import ChartPreviewStyles from '../ReactViews/Chart/chart-preview.scss';
+import ChartPreviewStyles from '../ReactViews/Custom/Chart/chart-preview.scss';
 
 const chartAttributes = [
     'src',
