+++ lib/ReactViews/Custom/Chart/chart-preview.scss
@@ -1,5 +1,5 @@
 @import '~terriajs-variables';
-@import '../../Sass/common/mixins';
+@import '../../../Sass/common/mixins';
 
 .chart {
   padding: 0;
