+++ lib/ReactViews/Custom/Chart/chart.scss
@@ -1,5 +1,5 @@
 @import '~terriajs-variables';
-@import '../../Sass/common/mixins';
+@import '../../../Sass/common/mixins';
 
 // D3 Axis
 .chart {
