+++ lib/ReactViews/Custom/Chart/downloadHrefWorker.js
@@ -1,8 +1,8 @@
 /* global onmessage:true */
 import defined from 'terriajs-cesium/Source/Core/defined';
 
-import DataUri from '../../Core/DataUri';
-import sortedIndices from '../../Core/sortedIndices';
+import DataUri from '../../../Core/DataUri';
+import sortedIndices from '../../../Core/sortedIndices';
 
 /**
  * Create combined arrays from arrays of column values, eg. [[values1, values2, values3], [values4, values5]].
